doc(readme): add note for IE users
At present there is a bug in the HTMLImports polyfill which only occurs
on IE. We have submitted a pull request to the team with the fix. In
the mean time, if you want to test on IE, you can work around the issue
by explicitly adding a script tag before you load system.js.